Update getting_values.py
The code should now be able to produce two csv files: one that contains the distance + velocity values at every 20 seconds, and code that says how far the rats are every second (note that it's not always at the same frame: for example, sometimes we might get the value at 2 1/2 seconds instead of 2, because frames might be missing. If we don't have data, for a second at all, the frame will be empty).